update linear feature weights proportionally
I've never heard this being discussed in literature - all the examples I
found have the weights being updated by the full amount, eg.
http://ai.berkeley.edu/reinforcement.html#Q8. This confused me, since
the overall update to the value (ie. across all features) would be far
too large. This proportional update is not unheard of, and in fact back
propagation goes further by adjusting based on the current weight.

In this case, experimentally, not having this term causes the weights to
blow up both negatively and positively towards infinity. Note that this
is different from divergence due to approximate off-policy learning (eg.
as discussed in 2nd edition of Sutton and Barto, Section 11.2). Although
the weights were exponentially large, they tend to cancel each other out
to give meaningful value estimates. The problem is that the values
eventually overflowed, leading to infinities and not-a-numbers.

Again, I am not sure why this is happening. It's possible that it's
because I'm using boolean (0 or 1) features instead of real numbers, and
that a different representation (say, -1 and 1) would solve this issue.
Alternately, it's also possible that simply reducing the learning rate
would fix this, since that's what this factor is doing numerically
(albeit dynamically based on the state, instead of making it constant or
gradually decreasing it over time).

Nonetheless, I'm surprised I've never seen this in the literature.